1. Created 2 new class(GlobalExceptionHandler and BookRequest) 2. Modified BookControllerIntegrationTest 3. Improving Business Logics

+++ src/main/java/com/example/bookstore/BookService.java
@@ -1,8 +1,13 @@
 package com.example.bookstore;
 
+import com.example.bookstore.model.Author;
 import com.example.bookstore.model.Book;
+import com.example.bookstore.repository.AuthorRepository;
 import com.example.bookstore.repository.BookRepository;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
+
 
 import java.util.List;
 import java.util.Optional;
@@ -11,9 +16,25 @@ import java.util.Optional;
 public class BookService {
 
     private final BookRepository bookRepository;
+    private final AuthorRepository authorRepository;
 
-    public BookService(BookRepository bookRepository) {
+    public BookService(BookRepository bookRepository, AuthorRepository authorRepository) {
         this.bookRepository = bookRepository;
+        this.authorRepository = authorRepository;
+    }
+
+    @Transactional
+    public Book save(Book book) {
+        // Checking if the author is in the database
+        if (book.getAuthor() != null && book.getAuthor().getId() != null) {
+            Author author = authorRepository.findById(book.getAuthor().getId())
+                    .orElseThrow(() -> new EntityNotFoundException("Author with ID " + book.getAuthor().getId() + " not found"));
+
+            book.setAuthor(author); // We tie the author to the book
+        }
+
+        // Saving a book to a database
+        return bookRepository.save(book);
     }
 
     public List<Book> getAll() {
@@ -24,11 +45,17 @@ public class BookService {
         return bookRepository.findById(id);
     }
 
-    public Book save(Book book) {
-        return bookRepository.save(book);
-    }
-
+    @Transactional
     public void delete(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Book with ID " + id + " not found");
+        }
         bookRepository.deleteById(id);
     }
+
+
+    public void deleteAll() {
+        bookRepository.deleteAll();
+    }
 }
+
